+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `s.dict_tag` as a convenience factory for building mapping specs for which
   the value spec tags are derived automatically from the corresponding dict keys (#52)
+- Add `s.rename` as a spec factory for validating that values are mapping types and
+  providing a default conformer to rename keys (#60)
 
 ## [v0.2.5] - 2020-04-10
 ### Added